Add --run-and-exit flag to asset_packaging example to be used in integration tests
Running the example with the flag will run the server in a separate goroutine and
then request / inline. The server will return a 503 if either the template parsing
fails or the template execution doesn't work correctly.
I also added some more features that are supported by Jet, namely blocks, inclusion,
imports as well as yielding.