add SetGlobal() to replicate the previous (buggy) behavior of setValue() which was exported via Set()
if setValue() couldn't find a variable with the given name in any of the scopes, it would create the variable in the top-most one instead of the current one. 6e9c5b9660f38b02c49513e3381f8da0e4b7854c fixed this, but outside code might depend on this bug to set global variables via Set(), so SetGlobal() can now be used to do that.