assign the range key/index in one-var ranges over map/slice/array (closes #158)
basically, if the ranger provides both a range value as well as an index, the one-var syntax sets that var to the index. if the ranger provides no index, one-var syntax will give you the value. this way, Jet's range now behaves like Go's range for maps, slices, arrays, and channels.