eval: Fix context when looping over unindexed ranger
This fixes a bug when looping over an unindex ranger (i.e. a ranger that
returns false for its ProvidesIndex() method).

Previously, the context for the range was not properly set, so it made
it impossible to use the context in such rangers.

This also fixes a panic when attempting to use an unindexed ranger
without a corresponding 'set' action.